pyproject.toml
[project]
name = "deploy-and-manage-adk-service"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "google-adk==1.3.0",
    "locust==2.37.10",
    "pg8000==1.31.2",
    "psycopg2-binary>=2.9.10",
    "python-dotenv==1.1.0",
]

[dependency-groups]
dev = [
    "pytest==8.4.0",
    "ruff==0.11.13",
]
